Split shell registration projection tests
